+++ database/migrations/2014_10_12_000002_create_products_table.php
@@ -1,26 +1,57 @@
 <?php
 
+use App\Models\Brand;
+use App\Models\Category;
+use App\Models\Product;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\Schema;
 
-return new class extends Migration
-{
+return new class extends Migration {
     /**
      * Run the migrations.
      */
     public function up(): void
     {
         Schema::create('products', function (Blueprint $table) {
+
             $table->id();
-            $table->string('product_title');
-            $table->unsignedBigInteger('category_id');
-            $table->foreign('category_id')->references('id')->on('categories');
-            $table->string('product_price');
-            $table->string('product_length');
-            $table->string('product_width');
+
+            $table->string('slug');
+
+            $table->string('title');
+
+            $table->string('description');
+
+            $table->string('thumbnail')
+                ->nullable();
+
+            $table->unsignedInteger('price')
+                ->default(0);
+
+
+            $table->foreignIdFor(Brand::class)
+                ->constrained()
+                ->cascadeOnUpdate()
+                ->cascadeOnDelete();
+
             $table->timestamps();
         });
+        Schema::create('category_product', function (Blueprint $table) {
+                $table->id();
+
+            $table->foreignIdFor(Category::class)
+                ->constrained()
+                ->cascadeOnUpdate()
+                ->cascadeOnDelete();
+
+            $table->foreignIdFor(Product::class)
+                ->constrained()
+                ->cascadeOnUpdate()
+                ->cascadeOnDelete();
+        });
+
+
     }
 
     /**
@@ -28,6 +59,10 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::dropIfExists('products');
+        if (app()->isLocal()) {
+            Schema::dropIfExists('products');
+            Schema::dropIfExists('category_product');
+
+        }
     }
 };
